fix: more cleanup, removed badge
