+++ fixed_io_test.go
